+++ magento/__init__.py
@@ -24,8 +24,19 @@ VISIBILITY_NOT_VISIBLE = 1
 ENABLED_PRODUCT = 1
 DISABLED_PRODUCT = 2
 
+# https://magento.stackexchange.com/a/244151/92968
+# https://github.com/magento/magento2/blob/c22a9db4/lib/internal/Magento/Framework/Image/Adapter/Gd2.php#L30-L34
+IMAGE_MIME_TYPES = {"image/gif", "image/jpeg", "image/png", "image/xbm", "image/wbmp"}
+
+DATE_ISO_8601_FORMAT = "%Y-%m-%d %H:%M:%S"
+
 
 def format_datetime(dt: datetime):
     """Format a datetime for Magento."""
     # "2021-07-02 13:19:18.300700" -> "2021-07-02 13:19:18"
     return dt.isoformat(sep=" ").split(".", 1)[0]
+
+
+def parse_datetime(s: str):
+    """Parse a datetime string from Magento."""
+    return datetime.strptime(s, DATE_ISO_8601_FORMAT)